msm8998: Change size of /vendor down to 1.0GB
OnePlus is doing their repartition in recent Open Beta, splitting the
1.4GB "last_parti" partition into 2 partitions: a 1.0GB sdf6 partition,
and a 0.5GB sdf7 "reserve3" partition (while "reserve1" and "reserve2"
are sda10 and sda11).

We used to assume the sdf6 is 1.4GB large. In this case, users may see a
"Invalid argument" failure while flashing the vendor image in recovery.
This patch decreased the expected image size down to 1.0GB, which is the
same as the stock ROM.

Change-Id: I11ea75a28c3310e50334423a052240f5db89bfd7